Check for outdated libs, ignoring Clojure(Script)
This commit ignores Clojure(Script) versions when comparing
dependencies.

This is response to the GitHub issue:
https://github.com/deps-app/versions/issues/8

Some users don't wish to have their versions of Clojure(Script) marked
as outdated, as they might intentionally wish to use that particular
version for their project. And the behaviour up until not was just to
look at the version number and find out if there is a later one, at
which point the project would be marked as having outdated dependency.